# Specialize FOR_ITER via GET_ITER