@sweeneyde 2022-04-06
Microbenchmarks show very little change to range iteration, but a decent (~1.13x) speedup to list iteration.
```
sween@DESKTOP-3GGN2FL:~/before_getiter2$ ./python -m pyperf timeit -s "R = range(10**5)" "for i in R: pass"
.....................
Mean +- std dev: 1.16 ms +- 0.03 ms
sween@DESKTOP-3GGN2FL:~/before_getiter2$ ./python -m pyperf timeit -s "R = list(range(10**5))" "for i in R: pass"
.....................
Mean +- std dev: 562 us +- 7 us


sween@DESKTOP-3GGN2FL:~/get_iter2$ ./python -m pyperf timeit -s "R = range(10**5)" "for i in R: pass"
.....................
Mean +- std dev: 1.15 ms +- 0.03 ms
sween@DESKTOP-3GGN2FL:~/get_iter2$ ./python -m pyperf timeit -s "R = list(range(10**5))" "for i in R: pass"
.....................
Mean +- std dev: 495 us +- 7 us
```