@sweeneyde 2022-04-06
Feedback/reviews/discussions are appreciated! I'm just not yet sure if this is a good approach at a high level.