@sweeneyde 2022-06-11
Closing this, as it is more complex and conceptually different from other specializations (type checks are cheap anyway).